Version 19/6/23

+++ app/src/main/java/com/gestordetrabajo/myapplication/List_Events_AdapterVH.java
@@ -1,20 +1,26 @@
 package com.gestordetrabajo.myapplication;
 
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.view.View;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.gestordetrabajo.myapplication.databinding.ActivityMainCalendarBinding;
 import com.gestordetrabajo.myapplication.db.MyDB;
 
 public class List_Events_AdapterVH extends RecyclerView.ViewHolder {
 
 
+
     TextView name;
     private List_Events_Adapter adapter;
     MyDB database;
+
+
 
     Context context;
     public List_Events_AdapterVH(@NonNull Context context,View itemView) {
@@ -26,12 +32,22 @@ public class List_Events_AdapterVH extends RecyclerView.ViewHolder {
 
         });
         itemView.findViewById(R.id.btn_eliminar_trabajo).setOnClickListener(view -> {
-            String id = adapter.items.get(getBindingAdapterPosition()).getId();
-            adapter.items.remove(getBindingAdapterPosition());
-            adapter.notifyItemRemoved(getBindingAdapterPosition());
             database = new MyDB(this.context);
-            database.deleteData(id);
+            AlertDialog.Builder dialog = new AlertDialog.Builder(itemView.getContext());
+            dialog.setTitle("Cuidado");
+            dialog.setMessage("Está seguro de eliminar el evento?");
+            dialog.setCancelable(false);
+            dialog.setIcon(R.drawable.baseline_delete_24);
+            dialog.setPositiveButton("Sí", (dialogInterface, i) -> {
+                String id = adapter.items.get(getBindingAdapterPosition()).getId();
+                adapter.items.remove(getBindingAdapterPosition());
+                adapter.notifyItemRemoved(getBindingAdapterPosition());
+                database.deleteData(id);
+            }).setNegativeButton("Cancelar", (dialogInterface, i) -> {
 
+            });
+            AlertDialog alertDialog = dialog.create();
+            alertDialog.show();
 
         });
     }
